// client/employee-management/src/app/module/hr-module/domain/showEmployeeTable.ts
export class ShowEmployeeTable {
  employeeId: number;
  name: string;
  cellPhone: string;
  alternatePhone: string;
  gender: string;
  dob: string;
  email: string;
  ssn: number;

  title: string;
  managerId: string;
  startDate: string;
  endDate: string;
  avatar: string;
  car: string;
  visaStatusId: string;
  visaStartDate: string;
  visaEndDate: string;
  driverLicense: string;
  driverLicenseExpDate: string;
  houseId: number;

  viewDetail: string;
}